Add Makefile, convert stdb:: to class
